Rename rounbound namespace to runeboundmagic
