+++ app/src/androidTest/java/com/example/runeboundmagic/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.rouneboundmagic
+package com.example.runeboundmagic
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.rouneboundmagic", appContext.packageName)
+        assertEquals("com.example.runeboundmagic", appContext.packageName)
     }
 }
