+++ app/src/main/AndroidManifest.xml
@@ -36,7 +36,7 @@
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
-                android:value="rouneboundmagic" />
+                android:value="runeboundmagic" />
         </activity>
     </application>
