+++ app/src/main/java/com/example/runeboundmagic/CrashHandlerApplication.kt
@@ -1,4 +1,4 @@
-package com.example.rouneboundmagic
+package com.example.runeboundmagic
 
 import android.app.Application
 import android.content.Intent
