+++ app/src/main/java/com/example/runeboundmagic/CrashReportActivity.kt
@@ -1,4 +1,4 @@
-package com.example.rouneboundmagic
+package com.example.runeboundmagic
 
 import android.content.ActivityNotFoundException
 import android.content.Intent
