+++ app/src/main/java/com/example/runeboundmagic/HeroOption.kt
@@ -1,4 +1,4 @@
-package com.example.rouneboundmagic
+package com.example.runeboundmagic
 
 import androidx.annotation.StringRes
